requirements.txt
# Core
sqlalchemy>=1.4.0
pyyaml>=6.0
python-dotenv>=0.19.0
aiofiles>=0.8.0
javalang>=0.13.0
requests>=2.31.0

# Tests
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Web dashboard (optional but recommended)
fastapi>=0.100.0
uvicorn>=0.20.0
markdown>=3.4.0
pygments>=2.16.0

# Optional (uncomment if needed)
# cx_oracle>=8.3.0            # Oracle driver
# jsqlparser>=0.1.0           # SQL parsing helper
# tree_sitter>=0.20.0         # Multi-language parsing
# antlr4-python3-runtime>=4.9.0
# openai>=1.0.0               # LLM integration (older)
openai>=1.35.0               # LLM integration (vLLM/OpenAI-compatible)
# langchain>=0.1.0
# faiss-cpu>=1.7.0
# sentence-transformers>=2.2.0
